Change float to int for Basestat

+++ src/main/kotlin/com/raycedni/PenAndPaperCompanion/gameSpecific/cyperpunkRED/stats/attributes/attributeClasses/Body.kt
@@ -3,4 +3,4 @@ package com.raycedni.PenAndPaperCompanion.gameSpecific.cyperpunkRED.stats.attrib
 import com.raycedni.PenAndPaperCompanion.gameSpecific.cyperpunkRED.stats.attributes.attributeClasses.AttributeBaseClass
 import com.raycedni.PenAndPaperCompanion.gameSpecific.cyperpunkRED.stats.attributes.attributeSkillListEnums.BodySkillListEnum
 
-class Body(override var points: Double) : AttributeBaseClass<BodySkillListEnum>("Body", points, enumValues())
+class Body(override var points: Int) : AttributeBaseClass<BodySkillListEnum>("Body", points, enumValues())
