+++ src/main/kotlin/com/raycedni/PenAndPaperCompanion/gameSpecific/cyperpunkRED/stats/attributes/attributeClasses/Intelligence.kt
@@ -2,5 +2,5 @@ package com.raycedni.PenAndPaperCompanion.gameSpecific.cyperpunkRED.stats.attrib
 
 import com.raycedni.PenAndPaperCompanion.gameSpecific.cyperpunkRED.stats.attributes.attributeSkillListEnums.IntelligenceSkillListEnum
 
-class Intelligence(override var points: Double) :
+class Intelligence(override var points: Int) :
     AttributeBaseClass<IntelligenceSkillListEnum>("Intelligence", points, enumValues())
