+++ src/main/kotlin/com/raycedni/PenAndPaperCompanion/gameSpecific/cyperpunkRED/stats/attributes/attributeClasses/Tech.kt
@@ -2,4 +2,4 @@ package com.raycedni.PenAndPaperCompanion.gameSpecific.cyperpunkRED.stats.attrib
 
 import com.raycedni.PenAndPaperCompanion.gameSpecific.cyperpunkRED.stats.attributes.attributeSkillListEnums.TechSkillListEnum
 
-class Tech(override var points: Double) : AttributeBaseClass<TechSkillListEnum>("Tech", points, enumValues())
+class Tech(override var points: Int) : AttributeBaseClass<TechSkillListEnum>("Tech", points, enumValues())
